+++ framework_src/native_library/FreSwiftExampleANE/FreSwift/FreSwiftMainController.swift
@@ -54,7 +54,7 @@ public extension FreSwiftMainController {
     /// - parameter value: value to trace to console
     /// - returns: Void
     func trace(_ value: Any...) {
-        var traceStr: String = "\(Self.TAG):"
+        var traceStr = "\(Self.TAG):"
         for i in 0..<value.count {
             traceStr = "\(traceStr) \(value[i]) "
         }
@@ -70,7 +70,7 @@ public extension FreSwiftMainController {
     /// - parameter value: value to trace to console
     /// - returns: Void
     func info(_ value: Any...) {
-        var traceStr: String = "\(Self.TAG):"
+        var traceStr = "\(Self.TAG):"
         for i in 0..<value.count {
             traceStr = "\(traceStr) \(value[i]) "
         }
@@ -85,7 +85,7 @@ public extension FreSwiftMainController {
     /// - parameter value: value to trace to console
     /// - returns: Void
     func warning(_ value: Any...) {
-        var traceStr: String = "\(Self.TAG):"
+        var traceStr = "\(Self.TAG):"
         for i in 0..<value.count {
             traceStr = "\(traceStr) \(value[i]) "
         }